requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosqlite
asyncpg
httpx
websockets
pytest
pytest-asyncio
